fixed: download link issue and %TMP% can now be cleaned up by the installer
Found a working download link that doesn't have an infinite redirect.
Updated
installer to use this file and removed references to the issue from
documentation.

This fix allowed the %TMP% to be cleaned up properly, so I uncommented
those lines of code in the installer.

Finally I added some feature requests in BUGS.txt.